apps/backend/container/requirements.txt
flask==3.0.0
gunicorn==21.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
pytesseract==0.3.10
pmtiles==3.2.0
pyvips==2.2.2
requests==2.31.0
